% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml_mm.R
\name{acml_mm}
\alias{acml_mm}
\title{Ascertainment Corrected Maximum Likelihood Estimator for the NSA design
The function estimates the parameters of interest for the NSA design using the ascertainment correct maximum likelihood (ACML)}
\usage{
acml_mm(
  mean.formula,
  lv.formula = NULL,
  t.formula = NULL,
  id,
  data,
  inits = NULL,
  samp.probs = c(1, 1, 1),
  offset = NULL,
  Q = 10,
  step.max = 1,
  step.tol = 1e-06,
  hess.eps = 1e-07,
  adapt.quad = FALSE,
  verbose = FALSE,
  maxiter = 100,
  return_args = FALSE
)
}
\arguments{
\item{mean.formula}{Formula for the marginal mean model in which a binary variable is regressed on a set of covariates. This is a required argument}

\item{lv.formula}{Formula for the latent part in the dependence model. This is optional and of the form ~ covariates}

\item{t.formula}{Formula for the transition part in the dependence model. This is optional and of the form ~ covariates}

\item{id}{Vector with \code{id} variable. This is a required argument}

\item{data}{Dataframe. This is a required argument}

\item{inits}{An optional list of length 3 containing initial values for marginal mean parameters and all dependence parameters. The format of the list should be: (1) estimates of the mean parameters, (2) estimates of the transition parameters (or NULL if only fitting a mLV model) and (3) estimates of the latent variable parameters (or NULL if only fitting a mT model). If NULL, initial values will be automatically generated. Default is set to NULL}

\item{samp.probs}{Vector of 3 numbers indicating the probability that a subject is sampled from the none
some and all stratum respectively. If not specified, the sampling probabilities are set to 1}

\item{offset}{an optional offset term.}

\item{Q}{A scalar to denote the number of quadrature points used for GH numerical integration. Default is 10.}

\item{step.max}{A scalar indicating the step maximum in the maximization algorithm. Default is 1}

\item{step.tol}{A scalar indicating the step tolerance in the maximization algorithm. Default is 1e-6}

\item{hess.eps}{A scalar indicating the perturbation value in the Hessian computation. Default is 1e-7}

\item{adapt.quad}{An indicator if adaptive quadrature is to be used.  NOT CURRENTLY IMPLEMENTED.}

\item{verbose}{An indicator of whether the steps of the maximization algorithm should be printed}

\item{maxiter}{A scalar indicating the maximum number of iteration in the maximization algorithm. Default is 100}

\item{return_args}{If TRUE returns the parameters used for the likelihood computation. Default is FALSE}
}
\value{
This function returns marginal mean (beta) and dependence parameters (alpha) along with the associated model and empirical covariance matrices.
}
\description{
Ascertainment Corrected Maximum Likelihood Estimator for the NSA design
The function estimates the parameters of interest for the NSA design using the ascertainment correct maximum likelihood (ACML)
}
\examples{
\donttest{
# perform the NSA design
design       <- nsa_design(Y = "Y", id = "id", n.sample = c(85, 100, 15), data = exampledat)
# Assume X is not collected for subjects not sampled
exampledat$X <- ifelse(exampledat$id \%in\% design$sampled.id, exampledat$X, NA)
# estimate the parameters using ACML
mod.acml    <- acml_mm(mean.formula = Y ~ X + time + Z + X:time, t.formula = ~1, id = id,
                       samp.probs = design$sample.probs, data = exampledat)
# print the results
summary(mod.acml)

}
}
