% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build-rmd.R
\name{bld_code_chunk}
\alias{bld_code_chunk}
\title{Build .Rmd r code chunk}
\usage{
bld_code_chunk(name = NULL, opts = NULL, body = NULL)
}
\arguments{
\item{name}{name to use for the chunk}

\item{opts}{options for code chunk}

\item{body}{content of the chunk}
}
\description{
Build a code chunk into character vector of lines to pass to \code{cat()} or
\code{writeLines()}.
}
\examples{
my_opts <- c("child=\\"my-child.Rmd", "fig.path=\\"figure/\\"")
 cat(bld_code_chunk("my-chunk", my_opts), sep = "\\n")
}
\keyword{internal}

