% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build-rmd.R
\name{bld_rmd_yaml}
\alias{bld_rmd_yaml}
\title{Build .Rmd yaml header}
\usage{
bld_rmd_yaml(keep_md = TRUE, variant = "markdown_github")
}
\arguments{
\item{keep_md}{Boolean for keeping markdown document when knitting}

\item{variant}{string of which markdown variant to knit.}
}
\description{
returns a character vector with all the necessary formatting to pass to
\code{cat()} or \code{writeLines()} and produce a yaml header for .Rmd
}
\examples{
cat(bld_rmd_yaml(), sep = "\\n")
}
\keyword{internal}

