% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build-rmd.R
\name{bld_yaml_line}
\alias{bld_yaml_line}
\title{Build a yaml parameter line}
\usage{
bld_yaml_line(param, value, indent_lvl, indent = "  ", sep = ":")
}
\arguments{
\item{param}{name of the parameter}

\item{value}{value of the parameter}

\item{indent_lvl}{how many levels of indentation, where each level is indent}

\item{indent}{was string to use for indentation. default is "  ".}

\item{sep}{what separates the param from the value. default is ":".}
}
\description{
Build a yaml header parameter line by pasting together indentation,
parameter name, separator, and value.
}
\examples{
bld_yaml_line("variant", "markdown_github", 2)
}
\keyword{internal}

