% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.TwoPhaseMM.R
\name{coef.TwoPhaseMM}
\alias{coef.TwoPhaseMM}
\title{Print the estimated coefficients}
\usage{
\method{coef}{TwoPhaseMM}(object, ...)
}
\arguments{
\item{object}{Object of class TwoPhaseMM}

\item{...}{other inputs}
}
\value{
Estimated coefficients from a TwoPhaseMM object
}
\description{
Print the estimated coefficients
}
