% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_functions.R
\name{DirectImputation}
\alias{DirectImputation}
\title{Impute the missing data and fit the complete data using the direct imputation method}
\usage{
DirectImputation(
  mean.formula,
  lv.formula,
  t.formula,
  id,
  data,
  sampled,
  X,
  stepmax = 1,
  samp.probs,
  Q = 10,
  M = 5,
  marg.exp.formula,
  m = 20,
  verbose = FALSE
)
}
\arguments{
\item{mean.formula}{Formula for the marginal mean model in which a binary variable is regressed on a set of covariates. This is a required argument}

\item{lv.formula}{Formula for the latent part in the dependence model. This is optional and of the form ~ covariates}

\item{t.formula}{Formula for the transition part in the dependence model. This is optional and of the form ~ covariates}

\item{id}{Column name for the id variable as recorded in the dataset. This is a required argument}

\item{data}{Full cohort data. This dataset should not have missing values for variables other than X. It should
contain columns named \code{id}, \code{X}, and \code{sampled}}

\item{sampled}{Column name for the sampling indicator variable as recorded in the dataset. This is a required argument}

\item{X}{Column name for the expensive exposure variable as recorded in the dataset. This is a required argument}

\item{samp.probs}{Vector of 3 numbers indicating the probability that a subject is sampled from the none
some and all stratum respectively. This is a required argument when \code{method} is set equal to indirect. It is not
used when \code{method} is set equal to direct}

\item{Q}{A scalar to denote the number of quadrature points used for GH numerical integration. Default is 10.}

\item{M}{Number of imputations. Default is 5.}

\item{marg.exp.formula}{Marginal exposure model of the form: X ~ Z where Z are the inexpensive covariates in \code{data}}

\item{m}{A scalar giving the number of iterations within each imputation. The default is 5.}

\item{verbise}{An indicator of whether the steps of the maximization algorithm should be printed. Default is set to FALSE}
}
\value{
The function returns the estimated coefficients and standard error from the direct imputation approach
}
\description{
Impute the missing data and fit the complete data using the direct imputation method
}
\keyword{internal}
