% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Compute the inverse of logit function}
\usage{
expit(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
The inverse logit of the input x
}
\description{
Compute the inverse of logit function
}
\keyword{internal}
