% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_functions.R
\name{FitImputedData}
\alias{FitImputedData}
\title{Fit imputed data using indirect imputation}
\usage{
FitImputedData(FIT, ImputedData, id, verbose = FALSE, iter.lim = 1000, q = 10)
}
\arguments{
\item{FIT}{Model fit for imputation}

\item{ImputedData}{List of datasets where the expensive exposure has been imputed}

\item{verbose}{An indicator of whether the steps of the maximization algorithm should be printed. Default is set to FALSE}

\item{iter.lim}{A scalar indicating the maximum number of iteration in the maximization algorithm. Default is 100}

\item{q}{A scalar to denote the number of quadrature points used for GH numerical integration. Default is 10.}
}
\value{
The function returns the estimated coefficients and standard error from the indirect imputation approach
}
\description{
Fit imputed data using indirect imputation
}
\keyword{internal}
