% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_functions.R
\name{ImputeDataInd}
\alias{ImputeDataInd}
\title{Create imputed dataset for the indirect imputation approach}
\usage{
ImputeDataInd(
  FIT,
  DAT,
  M = 10,
  ROB = FALSE,
  MARG.EXP.FORMULA,
  verbose = FALSE,
  EXPVAR = EXPVAR,
  Sampled = Sampled
)
}
\arguments{
\item{FIT}{TwoPhaseMM object}

\item{DAT}{complete dataset with sampled flag (Sampled)}

\item{M}{Total number of imputations. The default number is set to 10.}

\item{ROB}{Indicator for the variance estimation. If FALSE use model-based covariance, otherwise use sandwich estimator. Default is set to FALSE.}

\item{MARG.EXP.FORMULA}{marginal exposure model formula (X | Z). This is the imputation model without the offset}

\item{verbose}{Indicator. If yes print the number of imputed dataset. Default is set to FALSE.}

\item{EXPVAR}{exposure / expensive variable: Xe}

\item{Sampled}{should be set to 1 if included in the sample and 0 otherwise}
}
\value{
Return the imputed dataset
}
\description{
Create imputed dataset for the indirect imputation approach
}
\keyword{internal}
