% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{infer_next_inc}
\alias{infer_next_inc}
\title{Infer the next increment number}
\usage{
infer_next_inc(dames_df, what = c("folders", "files"), inc_by = 1)
}
\arguments{
\item{dames_df}{output from \code{read_dames}}

\item{what}{what to consider: files and/or folders.}

\item{inc_by}{the amount to increment by}
}
\description{
takes a df of tidy'd dames and determines the next increment to use.
}
\keyword{internal}

