% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/github-friendly-readmes.R
\name{knit_rmds_to_rdme}
\alias{knit_rmds_to_rdme}
\title{knit all rmarkdown files into .md files in their own directories and name each
README.md so as to take advantage of the github previewing.}
\usage{
knit_rmds_to_rdme()
}
\description{
knit all rmarkdown files into .md files in their own directories and name each
README.md so as to take advantage of the github previewing.
}
\keyword{int}

