% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/github-friendly-readmes.R
\name{knitme}
\alias{knitme}
\title{knit just the .rmd in the current working directory to readme.md}
\usage{
knitme()
}
\description{
knit just the .rmd in the current working directory to readme.md
}

