% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{left_pad}
\alias{left_pad}
\title{left pad inc if need be and convert to char}
\usage{
left_pad(inc)
}
\arguments{
\item{inc}{increment}
}
\description{
left pad inc if need be and convert to char
}
\keyword{internal}

