% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLScoreCalc.R
\name{LogLScoreCalc}
\alias{LogLScoreCalc}
\title{Compute the Log-Likelihood and the Score for Each Subject}
\usage{
LogLScoreCalc(
  params,
  subjectData,
  CondLike = FALSE,
  ParamLengths,
  EmpiricalCheeseCalc = FALSE,
  Q,
  W,
  Z,
  AdaptiveQuad = FALSE
)
}
\arguments{
\item{params}{Vector of parameters from the marginal mean and the dependence model parameters}

\item{subjectData}{List of lists with subject-specific data}

\item{CondLike}{TRUE/FALSE. Set to TRUE from ascertainment corrected likelihood}

\item{ParamLengths}{Vector of three elements: the number of parameters in the marginal mean model, the number of parameters for the transition component of the dependence model, the number of parameters for the dependence component of the dependence model}

\item{EmpiricalCheeseCalc}{TRUE/FALSE. Set to TRUE for robust standard errors estimation}

\item{Q}{A scalar to denote the number of quadrature points used for GH numerical integration. Default is 10.}

\item{W}{A scalar resulting from GH numerical integration}

\item{Z}{A scalar resulting from GH numerical integration}

\item{AdaptiveQuad}{TRUE/FALSE. Set to TRUE for adaptive quadrature}
}
\value{
Log-likelihood and ascertainment correction (if CondLike is set to be FALSE) computed for each subject in the data
}
\description{
Compute the Log-Likelihood and the Score for Each Subject
}
\keyword{internal}
