% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nd.R
\name{nd}
\alias{nd}
\title{create a new dame (nd)}
\usage{
nd(slug = "simple-slug", aoa = "AOA", path = ".", ptrn = "[_]",
  common = "../REF_00_template.Rmd")
}
\arguments{
\item{slug}{string to be describe slug-part of the file dame}

\item{aoa}{acronym or abbreviation to add to the file dame}

\item{path}{path to work from, defaults to current directory}

\item{ptrn}{this is the pattern to use for infering which files/folders are
part of the numbering and which are not.}

\item{common}{string path of a common template/reference/source file that
each dame imports as a child at the start of the document.}
}
\description{
describe the contents of a file in plain text, optionally adding an
acronym or abbreviation and \code{nd} will make a nice dame in its own
directory with left-padded numbering being infered based on what's already
there.
}
\examples{
\dontrun{
   b4 <- list.files(, recursive = T)
   nd("the first part of my project is to explore", "EDA")
   nd("the second part is to fit models", "MDL")
   nd("fit more models", "MDL")
   aft <- list.files(, recursive = T)
   setdiff(aft, b4)
   }
}

