% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{newdame}
\alias{newdame}
\title{assemble new dame based on provided arguments}
\usage{
newdame(slug = "my-next-file", aoa = "AOA", path = ".", ptrn = "[_]")
}
\description{
assemble new dame based on provided arguments
}
\keyword{internal}

