% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsa_design.R
\name{nsa_design}
\alias{nsa_design}
\title{Title
None-Some-All (NSA) Design}
\usage{
nsa_design(Y, id, n.sample, data = NULL)
}
\arguments{
\item{Y}{Column name with the \code{Y} variable. This is a required argument}

\item{id}{Column name with the \code{id} variable. This is a required argument}

\item{n.sample}{Vector for three values representing the number of people to be sampled from the none, some and all stratum respectively. This is a required argument}

\item{data}{Dataframe containing \code{Y} and \code{id}. This is a required argument}
}
\value{
A list with the following components:
\describe{
\item{sampled.id}{List of IDs number sampled in the none, some, all strata}
\item{sample.probs}{Vector of sampling probabilities from the none, some, all strata to be used in the analysis}
\item{n.stratum}{Vector specifying the number of subjects in the none, some, all strata}
\item{membership}{Vector specifying in which strata a subject belongs to (None, Some, All)}
}
}
\description{
Function to perform the NSA design. The function requires specification of the column name with the outcome, the column name with the ID,
the number of subjects the user wants to sample from the none, some and all strata and the dataframe. For each stratum, if the
user-specified number of subjects of being sampled is greater than the actual number of subjects, the function
will return every subjects from that stratum with a warning.
}
\examples{
\donttest{
data(exampledat)
nsa_design(Y = "Y", id = "id", n.sample = c(85, 100, 15), data = exampledat)
}

}
