% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_functions.R
\name{obs.loglik}
\alias{obs.loglik}
\title{Compute the observed data log-likelihood (Equation 2 in Tao et al 2017)}
\usage{
obs.loglik(
  X,
  Z = NULL,
  id,
  Bspline = NULL,
  n_sieve = NULL,
  thetas,
  sData.n1,
  sData.n2,
  theta.length,
  n1,
  m,
  Q,
  p,
  data.n1,
  data.n2
)
}
\arguments{
\item{X}{Character indicating the name of the expensive exposure in the dataset}

\item{Z}{Character indicating the name of the inexpensive covariates in the dataset}

\item{id}{Column name with the \code{id} variable.}

\item{Bspline}{Matrix of B-spline basis used for the sieve approximation of P(X|Z)}

\item{n_sieve}{Number of sieves used for the B-spline approximation of P(X|Z)}

\item{thetas}{Vector of parameters}

\item{sData.n1}{List of subject-specific data (includes only subjects who were not sample for expensive exposure ascertainment)}

\item{sData.n2}{List of subject-specific data (includes only subjects who were sample for expensive exposure ascertainment)}

\item{theta.length}{Length of the vector of parameters theta}

\item{n1}{Number of subjects who were not sampled for expensive exposure ascertainment}

\item{m}{Number of distinct values of the expensive exposure observed in phase two}

\item{Q}{Number of quadrature point to be used if there is a latent term in the dependence model}

\item{p}{p_{ij} used to approximate P(X|Z)}

\item{data.n1}{Dataset including only subjects not sampled for phase two}

\item{data.n2}{Dataset including only subjects sampled for phase two}
}
\value{
Return the observed data log-likelihood for the sample
}
\description{
Compute the observed data log-likelihood (Equation 2 in Tao et al 2017)
}
\keyword{internal}
