% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_functions.R
\name{Odds2Prob}
\alias{Odds2Prob}
\title{Transform the odds to a probability}
\usage{
Odds2Prob(odds)
}
\arguments{
\item{odds}{Odds}
}
\value{
Probability
}
\description{
Transform the odds to a probability
}
\keyword{internal}
