% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_functions.R
\name{pl.loglik}
\alias{pl.loglik}
\title{Compute the profile likelihood needed to estimate the coefficients' standard errors}
\usage{
pl.loglik(
  X,
  Z = NULL,
  Bspline = NULL,
  n_sieve = NULL,
  sData.n1,
  sData.n2,
  theta.length,
  id,
  data.n1,
  data.n2,
  thetas,
  p.est,
  n1,
  m,
  Q,
  maxiter = 100,
  tol = 1e-04
)
}
\arguments{
\item{X}{Character indicating the name of the expensive exposure in the dataset}

\item{Z}{Character indicating the name of the inexpensive covariates in the dataset}

\item{Bspline}{Matrix of B-spline basis used for the sieve approximation of P(X|Z)}

\item{n_sieve}{Number of sieves used for the B-spline approximation of P(X|Z)}

\item{sData.n1}{List of subject-specific data (includes only subjects who were not sample for expensive exposure ascertainment)}

\item{sData.n2}{List of subject-specific data (includes only subjects who were sample for expensive exposure ascertainment)}

\item{theta.length}{Length of the vector of parameters theta}

\item{id}{Character indicating the name of the subjects id variable in the dataset}

\item{data.n1}{Dataset including only subjects not sampled for phase two}

\item{data.n2}{Dataset including only subjects sampled for phase two}

\item{thetas}{Vector of parameters}

\item{p.est}{p_{ijk} estimated from the EM algorithm}

\item{n1}{number of subjects not sampled for expensive exposure ascertainment}

\item{m}{Number of distinct values of the expensive exposure observed in phase two}

\item{Q}{Number of quadrature points for the estimation of the random effect in the latent component of the dependence model.}

\item{maxiter}{Maximum number of iterations for the EM algorithm}

\item{tol}{Tolerance to decide on algorithm convergence}
}
\value{
Return the profile log-likelihood
}
\description{
Compute the profile likelihood needed to estimate the coefficients' standard errors
}
\keyword{internal}
