% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_functions.R
\name{prep_data}
\alias{prep_data}
\title{Prepare the data for the indirect imputation approach}
\usage{
prep_data(
  mean.formula,
  lv.formula,
  t.formula,
  id,
  data,
  inits,
  samp.probs,
  samp.probi,
  offset
)
}
\arguments{
\item{mean.formula}{Formula for the marginal mean model in which a binary variable is regressed on a set of covariates.}

\item{lv.formula}{Formula for the latent part in the dependence model. This is optional and of the form ~ covariates}

\item{t.formula}{Formula for the transition part in the dependence model. This is optional and of the form ~ covariates}

\item{id}{Vector of \code{id} variable. This is a required argument}

\item{data}{Dataframe. This is a required argument}

\item{inits}{An optional list of length 3 containing initial values for marginal mean parameters and all dependence parameters. The format of the list should be: (1) estimates of the mean parameters, (2) estimates of the transition parameters (or NULL if only fitting a mLV model) and (3) estimates of the latent variable parameters (or NULL if only fitting a mT model). If NULL, initial values will be automatically generated. Default is set to NULL}

\item{samp.probs}{Vector of 3 numbers indicating the probability that a subject is sampled from the none
some and all stratum respectively. If not specified, the sampling probabilities are set to 1}

\item{samp.probi}{Vector of sampling probabilities for each individual}

\item{offset}{an optional offset term.}
}
\description{
Prepare the data for the indirect imputation approach
}
\keyword{internal}
