% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.TwoPhaseMM.R
\name{print.summary.TwoPhaseMM}
\alias{print.summary.TwoPhaseMM}
\title{Title}
\usage{
\method{print}{summary.TwoPhaseMM}(x, ...)
}
\arguments{
\item{x}{object of class TwoPhaseMM}

\item{...}{additional inputs}
}
\description{
Title
}
