% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_functions.R
\name{ptheta.fun}
\alias{ptheta.fun}
\title{Compute the Log-Likelihood for a marginalized transition and latent variable model}
\usage{
ptheta.fun(sData, params, p.length, X, m, Q, d)
}
\arguments{
\item{sData}{List of subject-specific data}

\item{params}{Vector of model parameters}

\item{p.length}{Length of the vector of model parameters}

\item{X}{Character indicating the name of the expensive exposure in the dataset}

\item{m}{Number of distinct values of the expensive exposure observed in phase two}

\item{Q}{Number of quadrature point to be used if there is a latent term in the dependence model}

\item{d}{Dataset}
}
\value{
Returns the log-likelihood computed for each subject
}
\description{
Compute the Log-Likelihood for a marginalized transition and latent variable model
}
\keyword{internal}
