% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_functions.R
\name{ptheta.input}
\alias{ptheta.input}
\title{Create Subject-Specific Data to Use in the SMLE function}
\usage{
ptheta.input(
  Y,
  X,
  mean.formula,
  t.formula,
  lv.formula,
  id,
  m,
  weights = NULL,
  d
)
}
\arguments{
\item{Y}{Character indicating the name of the outcome variable in the dataset}

\item{X}{Character indicating the name of the expensive exposure in the dataset}

\item{mean.formula}{Formula for the marginal mean model in which a binary variable is regressed on a set of covariates.}

\item{t.formula}{Formula for the transition part in the dependence model. This is of the form ~ covariates}

\item{lv.formula}{Formula for the latent part in the dependence model. This is of the form ~ covariates}

\item{id}{Column name with the \code{id} variable.}

\item{m}{Number of distinct values of the expensive exposure observed in phase two}

\item{weights}{Vector of weights for the M-step in the EM-algorithm}

\item{d}{dataset}
}
\value{
Returns a list of data per each subject to be used in the likelihood computation
}
\description{
Create Subject-Specific Data to Use in the SMLE function
}
\keyword{internal}
