% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_dames}
\alias{read_dames}
\title{read_dames}
\usage{
read_dames(dames, var_sep = "_", name_vars = c("aoa", "inc", "slug", "ext"))
}
\arguments{
\item{dames}{stands for datafied names}

\item{var_sep}{}

\item{name_vars}{}
}
\description{
take files names that are arranged in data_categories and turn them into a
tidy data frame. Defaults to 4 variables: abbreviation or acronym /
increment / slug / extension.
}
\keyword{internal}

