% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_design.R
\name{res_design}
\alias{res_design}
\title{Residual-Based Design}
\usage{
res_design(
  Y,
  id,
  mean.formula,
  t.formula = NULL,
  lv.formula = NULL,
  n.sample,
  statistic = "mean",
  data = NULL
)
}
\arguments{
\item{Y}{Column name for outcome variable as recorded in the dataset. This is a required argument}

\item{id}{Column name for the \code{id} variable as recorded in the dataset. This is a required argument}

\item{mean.formula}{Formula for the marginal mean model in which a binary variable is regressed on a set of covariates. This is a required argument}

\item{t.formula}{Formula for the transition part in the dependence model. This is optional and of the form ~ covariates}

\item{lv.formula}{Formula for the transition part in the dependence model. This is optional and of the form ~ covariates}

\item{n.sample}{Number of people to be sampled with the residual-based design. This is a required argument}

\item{statistic}{Name of the statistic to be used for the residual-based design. Current statistics supported are \code{"mean"}, \code{"abs.mean"}, \code{"Variance"}. Default is \code{"mean"}}

\item{data}{Dataframe containing \code{Y} and \code{id}. This is a required argument}
}
\value{
A list with the following components:
\describe{
\item{sampled.id}{List of IDs number sampled}
}
}
\description{
Function to perform residual-based designs. The function requires specification of the column name with the outcome, the column name with the ID,
the marginal mean model, at least one between the transition component or the latent component of the dependence model,
the number of subjects the user wants to sample, and the dataframe.
}
\examples{
\donttest{
data(exampledat)
res_design(Y = "Y", id = "id", mean.formula = Y ~ Z + time, t.formula = ~1, lv.formula = NULL,
n.sample = 200, statistic = "abs.mean", data = exampledat)
}
}
