% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_functions.R
\name{smle_mm}
\alias{smle_mm}
\title{Sieve Maximum Likelihood Estimator (SMLE) for Two-Phase Designs and Marginalized Transition and Latent Variable Models}
\usage{
smle_mm(
  Y,
  X,
  Z = NULL,
  Time,
  Bspline = NULL,
  n_sieve = NULL,
  mean.formula = NULL,
  t.formula = NULL,
  lv.formula = NULL,
  id,
  data,
  Q = 10,
  maxiter = 100,
  tol = 1e-04,
  no_SE = TRUE,
  verbose = FALSE,
  hnscale = 1
)
}
\arguments{
\item{Y}{Column name for outcome variable as recorded in the dataset. This is a required argument}

\item{X}{Column name for expensive exposure variable as recorded in the dataset. This is a required argument}

\item{Z}{Vector of columns name for inexpensive variables as recorded in the dataset. This is an optional argument}

\item{Time}{Column name for the time variable as recorded in the dataset. This is a required argument}

\item{Bspline}{Matrix with B-spline basis. This is a required argument.}

\item{n_sieve}{Number of sieves for the B-spline approximation}

\item{mean.formula}{Formula for the marginal mean model in which a binary variable is regressed on a set of covariates. This is a required argument}

\item{t.formula}{Formula for the transition part in the dependence model. This is optional and of the form ~ covariates}

\item{lv.formula}{Formula for the latent part in the dependence model. This is optional and of the form ~ covariates}

\item{id}{Column name for the \code{id} variable as recorded in the dataset. This is a required argument}

\item{data}{Dataframe containing \code{Y}, \code{id}, \code{Z} and \code{X}. This is a required argument}

\item{Q}{Number of quadrature points for the estimation of the random effect in the latent component of the dependence model. Default is set to 10}

\item{maxiter}{Maximum number of iterations for the SMLE algorithm. Maximum is set to 100. This is an optional argument}

\item{tol}{Tolerance for the convergence criterion in the EM algorithm. The default value is \code{1E-4}. This is an optional argument}

\item{no_SE}{If \code{TRUE}, the variances of the parameter estimators will not be estimated. The default value is \code{FALSE}. This is an optional argument}

\item{verbose}{If \code{TRUE}, show estimated coefficients at each step of the EM algorithm. The default value is \code{FALSE}. This is an optional argument}

\item{hnscale}{Specifies the scale of the perturbation constant in the variance estimation. Default is set to 1. This is an optional argument}
}
\value{
This function returns marginal mean (beta) and dependence parameters (alpha) along with the associated model and empirical covariance matrices.
}
\description{
Estimate the parameters of interest
}
\examples{
\donttest{
data(exampledat)
# perform the NSA design
design       <- nsa_design(Y = "Y", id = "id", n.sample = c(85, 100, 15), data = exampledat)
# Assume X is not collected for subjects not sampled
exampledat$X <- ifelse(exampledat$id \%in\% design$sampled.id, exampledat$X, NA)
# set up P(X|Z)
Bspline_Z <- cbind(as.numeric(exampledat$Z == 0), as.numeric(exampledat$Z == 1))
n_sieve             <- ncol(Bspline_Z)
colnames(Bspline_Z) <- paste("bs", 1:n_sieve, sep="")
exampledat          <- cbind(exampledat, Bspline_Z)
# estimate the parameters using SMLE
mod.smle <- smle_mm(mean.formula = Y ~ X + time + Z + X:time, t.formula = ~1, lv.formula = NULL,
Y = "Y", X = "X", Z = "Z", Time = "time",
id = "id", Bspline = Bspline_Z, n_sieve = n_sieve, no_SE = FALSE,
data = exampledat)
summary(mod.smle)
}
}
