% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TwoPhaseMM.R
\name{summary.TwoPhaseMM}
\alias{summary.TwoPhaseMM}
\title{Summarise the results}
\usage{
\method{summary}{TwoPhaseMM}(object, ...)
}
\arguments{
\item{object}{Object of class TwoPhaseMM}

\item{...}{Additional input}
}
\value{
The function results result summaries of a TwoPhaseMM object
}
\description{
Summarise the results
}
