% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.TwoPhaseMM.R
\name{vcov.TwoPhaseMM}
\alias{vcov.TwoPhaseMM}
\title{Print Covariance Matrix}
\usage{
\method{vcov}{TwoPhaseMM}(object, ...)
}
\arguments{
\item{object}{Object of class TwoPhaseMM}

\item{...}{additional input}
}
\value{
Return a list with two elements: the covariance matrix of the coefficients in the
marginal mean model and the covariance matrix of the coefficients in the dependence model
}
\description{
Print Covariance Matrix
}
