% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wee_mm.R
\name{wee_mm}
\alias{wee_mm}
\title{Weighted Estimating Equations
The function estimates the parameters of interest using weighted estimating equations}
\usage{
wee_mm(
  mean.formula,
  lv.formula = NULL,
  t.formula = NULL,
  id,
  data,
  inits = NULL,
  weights = NULL,
  offset = NULL,
  Q = 10,
  step.max = 1,
  step.tol = 1e-06,
  hess.eps = 1e-07,
  adapt.quad = FALSE,
  verbose = FALSE,
  maxiter = 100,
  return_args = FALSE
)
}
\arguments{
\item{mean.formula}{Formula for the marginal mean model in which a binary variable is regressed on a set of covariates. This is a required argument}

\item{lv.formula}{Formula for the latent part in the dependence model. This is optional and of the form ~ covariates}

\item{t.formula}{Formula for the transition part in the dependence model. This is optional and of the form ~ covariates}

\item{id}{Column name with the \code{id} variable. This is a required argument}

\item{data}{Dataframe containing \code{Y} and \code{id}. This is a required argument}

\item{inits}{An optional list of length 3 containing initial values for marginal mean parameters and all dependence parameters. The format of the list should be: (1) estimates of the mean parameters, (2) estimates of the transition parameters (or NULL if only fitting a mLV model) and (3) estimates of the latent variable parameters (or NULL if only fitting a mT model). If NULL, initial values will be automatically generated. Default is NULL}

\item{weights}{A vector of length nrow(dat) indicating the sampling weights. This is a required argument}

\item{offset}{an optional offset term. NOT CURRENTLY IMPLEMENTED}

\item{Q}{A scalar to denote the number of quadrature points used for GH numerical integration. Default is 10.}

\item{step.max}{A scalar indicating the step maximum in the maximization algorithm. Default is 1}

\item{step.tol}{A scalar indicating the step tolerance in the maximization algorithm. Default is 1e-6}

\item{hess.eps}{A scalar indicating the perturbation value in the Hessian computation. Default is 1e-7}

\item{adapt.quad}{An indicator if adaptive quadrature is to be used.  NOT CURRENTLY IMPLEMENTED.}

\item{verbose}{An indicator of whether the steps of the maximization algorithm should be printed}

\item{maxiter}{A scalar indicating the maximum number of iteration in the maximization algorithm. Default is 100}

\item{return_args}{If TRUE returns the parameters used for the likelihood computation. Default is FALSE}
}
\value{
This function returns marginal mean (beta) and dependence parameters (alpha) along with the associated model and empirical covariance matrices.
}
\description{
Weighted Estimating Equations
The function estimates the parameters of interest using weighted estimating equations
}
\examples{
\dontrun{
# get the sampling probabilities from a two phase design
samp.probs <- ifelse(design$membership == "None", design$sample.probs[1],
      ifelse(design$membership == "Some", design$sample.probs[2], design$sample.probs[3]))
# estimate the parameters using WEE
mod.wee    <- wee_mm(mean.formula = Y ~ X + time + Z + X:time, t.formula = ~1,
                    id = id, weights = 1/samp.probs, data = exampledat)
summary(mod.wee)
}

}
